refactor!: restructure project as proper Python package
- Move code into supabase_mcp package directory
- Add proper entry points for server and inspector
- Configure development tools (ruff, mypy)
- Update project metadata and build configuration
- Add graceful shutdown handling

BREAKING CHANGE: Project structure changed from src/ to supabase_mcp/
Users need to reinstall with `uv sync` (package=true enabled)
