+++ CHANGELOG.MD
@@ -1,0 +1,26 @@
+# Changelog
+
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
+
+## [Unreleased]
+### Added
+- Full support for native Python SDK
+- Added deploy config for Smithery.ai
+- 1-command installation via pipx
+
+### Changed
+- Improved SQL syntax validation (sqlparse)
+- Improved setup instructions in README
+
+
+## [0.1.0] - 2025-02-16
+### Added
+- Initial release
+- Basic MCP server functionality
+- Supabase database connection support
+- Integration with Cursor and Windsurf IDEs
+
+[Unreleased]: https://github.com/alexander-zuev/supabase-mcp-server/compare/v0.1.0...HEAD
+[0.1.0]: https://github.com/alexander-zuev/supabase-mcp-server/releases/tag/v0.1.0
