+++ smithery.yaml
@@ -11,10 +11,10 @@ startCommand:
     properties:
       supabaseProjectRef:
         type: string
-        description: The project reference for the Supabase project.
+        description: The project reference of Supabase project you want to connect to.
       supabaseDbPassword:
         type: string
-        description: The database password for the Supabase project.
+        description: The database password of Supabase project you want to connect to.
   commandFunction:
     # A function that produces the CLI command to start the MCP on stdio.
     |-
