+++ supabase_mcp/settings.py
